brain/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
edge-tts==6.1.9
google-generativeai==0.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
piper-tts